requirements.txt
numpy>=1.20.1
pandas>=1.1.3
dataframe-image>=0.1.1
lxml>=4.6.4
tqdm~=4.54.1
python-telegram-bot>=13.8.1
tinvest>=3.0.5
schedule~=1.1.0